+++ authoring-api/lib/BaseREST.js
@@ -174,7 +174,8 @@ class BaseREST {
                 }
 
                 // Add a log warning for the retry.
-                if (retVal) {
+                const attempt = response ? response.attempts : ((err && err.attempts) ? err.attempts : 1);
+                if (retVal && (attempt < maxAttempts)) {
                     utils.logWarnings(context, i18n.__("retry_failed_request", {id: requestOptions.instanceId, message: error.original_message || error.log}));
                 }
             }